Create class to handle rental methods
